Daemonization of tyroneBot
tyroneBot can now be run as a daemon. This means that it can be
interacted with like httpd/apache2/nginx/sshd via the 'system' command.
It uses the pid /var/run/tyrone.pid to allow for access to the daemon
runtime environment. Logging is also now supported under
/var/log/tyrone.log.

New command: "I think Tyrone does" prints out a sassy response
Tyrone now logs any message with his name.
